+++ frontend/packages/ui-kit/src/card.tsx
@@ -1,27 +1,16 @@
-import { type JSX } from "react";
+import * as React from "react";
 
-export function Card({
-  className,
-  title,
-  children,
-  href,
-}: {
-  className?: string;
-  title: string;
+export interface CardProps extends React.HTMLAttributes<HTMLDivElement> {
   children: React.ReactNode;
-  href: string;
-}): JSX.Element {
+}
+
+export function Card({ className, children, ...props }: CardProps) {
   return (
-    <a
-      className={className}
-      href={`${href}?utm_source=create-turbo&utm_medium=basic&utm_campaign=create-turbo"`}
-      rel="noopener noreferrer"
-      target="_blank"
+    <div
+      className={`bg-white rounded-lg border border-gray-200 shadow-sm ${className}`}
+      {...props}
     >
-      <h2>
-        {title} <span>-&gt;</span>
-      </h2>
-      <p>{children}</p>
-    </a>
+      {children}
+    </div>
   );
 }